+++ src/Metadata/Identity/IdentityExpression.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JacyImp\ApiPlatformRateLimiter\Metadata\Identity;
+
+interface IdentityExpression
+{
+}